medicare-backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
alembic==1.13.1
psycopg2-binary==2.9.9
pydantic==2.7.1
pydantic-settings==2.2.1
python-jose[cryptography]==3.3.0
httpx==0.27.0
celery==5.3.6
redis==5.0.4
reportlab==4.2.0
mlxtend==0.23.1
tensorflow-cpu==2.16.2
numpy==1.26.4
pandas==2.0.3
scikit-learn==1.4.2
python-multipart==0.0.9
